requirements.txt
gradio==3.19.1
openai
python-dotenv
fastapi
uvicorn[standard]
langchain